Finished refactor of mdi / tabbed document architecture.
