+++ packages/app/client/src/ui/shell/mdi/tabbedDocument/tabbedDocument.tsx
@@ -58,12 +58,3 @@ export class TabbedDocument extends React.Component<TabbedDocumentProps> {
     );
   }
 }
-
-export const Tab = props => props.children;
-export const Content = props =>
-  <TabbedDocumentContentWrapper
-    documentId={ props.documentId }
-    hidden={ props.hidden }
-  >
-    { props.children }
-  </TabbedDocumentContentWrapper>;
